Add baseline Discord Bot implementation
I apologise for the atrociously large commit... In any case, the Bot works so far as to be able to take users from joining, through email verification using Spam, to assigning the Verified role to the user if verification succeeds. Look at the TODOs for what more is to be done.